# BackEnd/API/openapi.yaml
openapi: 3.0.3
info:
  title: OmniSource Backend API
  version: "2.0"
  description: |
    OmniSource 后端总接口文档。HTTP 接口基于 Spring Boot 3，默认服务端口 8081。
    WebSocket 通道通过 x-websocket 扩展字段记录，标准 OpenAPI paths 仅描述 HTTP/SSE 接口。
servers:
  - url: http://localhost:8081
    description: Local development server
...
paths:
  /api/aigc/chat:
    get:
      tags: [AIGC]
      summary: 普通 AI 问答
      parameters:
        - $ref: "#/components/parameters/MessageQuery"
      responses:
        "200":
          description: 纯文本回答
          content:
            text/plain:
              schema:
                type: string
              example: 你好，我是 OmniSource AI 助手。
    post:
      tags: [AIGC]
      summary: 普通 AI 问答（JSON，小程序推荐）
      description: 请求体为 JSON，响应为统一 Result，data 为模型回复字符串。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  description: 用户问题
            example:
              message: 介绍一下福州寿山石雕
      responses:
        "200":
          description: 统一 JSON，data 为回答文本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringResult"
              example:
                code: 200
                message: success
                data: 寿山石雕是福州代表性传统工艺之一……
                timestamp: "2026-05-03T10:00:00"
  /api/aigc/stream:
    get:
      tags: [AIGC]
      summary: SSE 流式 AI 问答
      description: 内部会拼接 RAG 检索上下文后流式输出。
      parameters:
        - name: message
          in: query
          required: false
          schema:
            type: string
            default: 你好，请自我介绍
      responses:
        "200":
          description: SSE 文本流
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                data: 寿山石雕
                data: 以石质温润
                data: 和雕刻精细见长
  /api/aigc/multimodal:
    post:
      tags: [AIGC]
      summary: 图片理解问答
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultimodalRequest"
            example:
              imageUrl: https://example.com/image.png
              question: 请描述这张图片
      responses:
        "200":
          description: 图片理解结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StringResult"
              example:
                code: 200
                message: success
                data: 图片中展示了一件传统工艺作品，具有雕刻纹理和装饰性细节。
                timestamp: "2026-05-03T10:00:00"
  /api/aigc/image:
    post:
      tags: [AIGC]
      summary: 生成图片
      description: 传入 roomId 和有效 agentCode 时，生成结果会写入聊天室并广播。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageGenerateRequest"
      responses:
        "200":
          description: 图片生成结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerateResult"
              example:
                code: 200
                message: success
                data:
                  imageUrl: https://java-ai-fzu.oss-cn-beijing.aliyuncs.com/OmniSource/chatroom/example.png
                  agentCode: fz_shoushan_stone
                  agentName: 寿山石雕器灵
                  agentAvatar: https://example.com/avatar.png
                  sourceImageUrl: https://java-ai-fzu.oss-cn-beijing.aliyuncs.com/OmniSource/source/福州/寿山石雕.png
                  messageId: "101"
                timestamp: "2026-05-03T10:00:00"
  /api/aigc/image/tasks/{taskId}:
    get:
      tags: [AIGC]
      summary: 查询图片生成任务
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 图片生成任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerationTaskResult"
              example:
                code: 200
                message: success
                data:
                  id: 1
                  taskId: task_abc123
                  userId: 1
                  roomId: 1001
                  prompt: 生成一张寿山石雕风格海报
                  style: 国风插画
                  status: SUCCEEDED
                  resultUrl: https://example.com/generated.png
                  errorMessage: null
                  progress: 100
                  model: qwen-image-2.0-pro
                  createTime: "2026-05-03T10:00:00"
                  updateTime: "2026-05-03T10:00:30"
                timestamp: "2026-05-03T10:00:30"
...
  /api/chat:
    post:
      tags: [Chat]
      summary: 发起多 Agent 联合问答
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultiAgentChatRequest"
      responses:
        "200":
          description: 多 Agent 问答结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiAgentChatResult"
              example:
                code: 200
                message: 多智能体问答成功
                data:
                  sessionId: session_001
                  query: 寿山石雕有哪些特点？
                  heritageId: fz_shoushan_stone
                  finalAnswer: 寿山石雕的特点包括石质温润、色彩丰富、因材施艺和题材多样。
                  agentReplies:
                    - agentCode: fz_shoushan_stone
                      title: 寿山石雕器灵
                      content: 寿山石雕重视石色、纹理与雕刻题材的结合。
                      references:
                        - chunk-001
                      searchUsed: false
                  retrievals:
                    - id: chunk-001
                      title: 寿山石雕
                      score: 0.92
                      content: 寿山石雕是福州传统工艺。
                      metadata:
                        city: 福州
                  webSearchResult: null
                timestamp: "2026-05-03T10:00:00"
  /api/chat/{sessionId}:
    get:
      tags: [Chat]
      summary: 查询会话最近一次结果
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 会话结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiAgentChatResult"
              example:
                code: 200
                message: success
                data:
                  sessionId: session_001
                  query: 寿山石雕有哪些特点？
                  heritageId: fz_shoushan_stone
                  finalAnswer: 寿山石雕的特点包括石质温润、色彩丰富、因材施艺和题材多样。
                  agentReplies: []
                  retrievals: []
                  webSearchResult: null
                timestamp: "2026-05-03T10:00:00"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    MessageQuery:
      name: message
      in: query
      required: false
      schema:
        type: string
        default: 你好
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
  schemas:
    ResultBase:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        timestamp:
          type: string
          format: date-time
    ErrorResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              nullable: true
      example:
        code: 400
        message: 参数错误
        data: null
        timestamp: "2026-05-03T10:00:00"
...
    StringResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              type: string
...
    MultiAgentChatResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MultiAgentChatResponse"
...
    ImageGenerateResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                imageUrl:
                  type: string
                  format: uri
                agentCode:
                  type: string
                agentName:
                  type: string
                agentAvatar:
                  type: string
                sourceImageUrl:
                  type: string
                messageId:
                  type: string
    ImageGenerationTaskResult:
      allOf:
        - $ref: "#/components/schemas/ResultBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImageGenerationTask"
...
    MultimodalRequest:
      type: object
      properties:
        imageUrl:
          type: string
          format: uri
        question:
          type: string
          default: 请描述这张图片
      required: [imageUrl]
    ImageGenerateRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          default: 1
        roomId:
          type: integer
          format: int64
          nullable: true
        prompt:
          type: string
        style:
          type: string
        agentCode:
          type: string
      required: [prompt]
      example:
        userId: 1
        roomId: 1001
        prompt: 生成一张寿山石雕风格海报
        style: 国风插画
        agentCode: fz_shoushan_stone
    MultiAgentChatRequest:
      type: object
      properties:
        sessionId:
          type: string
        query:
          type: string
        heritageId:
          type: string
        topK:
          type: integer
          default: 3
        searchEnabled:
          type: boolean
        agentCodes:
          type: array
          items:
            type: string
      required: [query]
      example:
        sessionId: optional-session-id
        query: 寿山石雕有哪些特点？
        heritageId: fz_shoushan_stone
        topK: 3
        searchEnabled: true
        agentCodes:
          - fz_shoushan_stone
          - fz_lacquerware
...
    MultiAgentChatResponse:
      type: object
      properties:
        sessionId:
          type: string
        query:
          type: string
        heritageId:
          type: string
        finalAnswer:
          type: string
        agentReplies:
          type: array
          items:
            $ref: "#/components/schemas/AgentReplyResponse"
        retrievals:
          type: array
          items:
            $ref: "#/components/schemas/RagRetrievalResponse"
        webSearchResult:
          type: string
    AgentReplyResponse:
      type: object
      properties:
        agentCode:
          type: string
        title:
          type: string
        content:
          type: string
        references:
          type: array
          items:
            type: string
        searchUsed:
          type: boolean
    RagRetrievalResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        score:
          type: number
          format: double
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
    ImageGenerationTask:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: string
        userId:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
        prompt:
          type: string
        style:
          type: string
        status:
          type: string
        resultUrl:
          type: string
        errorMessage:
          type: string
        progress:
          type: integer
        model:
          type: string
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
...